Add vertex colors to shaders, support drawing with color
Allow DrawImage on ebiten.Image to take an optional color, which
ultimately ends up dumped into a vertex buffer, allowing drawing the
same image in many colors without needing to reset the GL state.

This probably ought to be cleaner or prettier, but appears to work.

In restorable/images_test.go, vertices() is no longer used, so
it's been removed, and the internal DrawImage calls are adjusted not
to use tints.