Initial rough-draft benchmarking utility
This isn't great, because it's hard to get samples from the GL rendering,
which happens in a far-away place. (It might be possible to hack something
in, by providing a SetBenchmark(b *testing.B) on a GraphicsContext or
something which can StartTimer/StopTimer for the actual rendering as well.

But it's anything.