rework benchmark to be saner
Instead of starting/stopping benchmark in two different places,
just start/stop it for the whole update function in graphicscontext.go,
which is both (1) where the actual render happens, (2) where the
user-provided update function is called. This improves the utility
of the benchmarks slightly.

Switch to using an interface instead of "testing.B", so you could
in theory provide your own benchmark functions, but also the testing
package doesn't have to be included in things other than the benchmark
example.

Drop hundreds of lines of unused test cases from the file this
was originally copied from.

Also, fake up command-line arguments so a sane minimum benchmark time
is used, because the full-second default of the benchmark code is way too
large here.