first-pass attempt at a color feature
This is a pretty awkward interface, mostly because of all the abstraction
layers between the user-visible ebiten.Image and the actual drawing
code.

The idea is to change the actual draw calls used for
internal/graphics.Image to drop things in the command queue, adding
an optional slice of colors. If only one color is in the slice, that
color is used for all four vertexes, but if you provide a four-color
slice, then they're used as the four vertexes. If the slice is nil,
we use the identity color.

This design is flawed in a number of ways, not least of which is
that color.RGBA is integer 0-255 values, and openGL wants float 0-1
values.

But! It does establish that displaying things with four colors
for their vertexes is not measurably slower than drawing them
without, and maybe 1% slower than not having the color slots in
the vertex array at all, if that... But it's hundreds of times
faster than setting colors using a colorM.